requirements.txt
gradio
langchain
PyYAML
PyMuPDF
transformers
smolagents